+++ src/redux/common/slice.ts
@@ -1,38 +1,14 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { IAddressData, IRelatedAddress } from '../../interface/address.interface';
+import { IAddressData } from '../../interface/address.interface';
 
 export interface CounterState {
     addressList: IAddressData[]
-    selectAddress : IRelatedAddress
     initialized : boolean
 }
 
-export const defaultSelectAddress = {
-  address: '',
-  address_name: '',
-  address_type: '',
-  road_address: {
-    address_name: '',
-    building_name: '',
-    main_building_no: '',
-    region_1depth_name: '',
-    region_2depth_name: '',
-    region_3depth_name: '',
-    road_name: '',
-    sub_building_no: '',
-    underground_yn: '',
-    x: '',
-    y: '',
-    zone_no: '',
-  },
-  x: 126.978652258309,
-  y: 37.566826004661,
-};
-
 const initialState: CounterState = {
   addressList: [],
-  selectAddress: defaultSelectAddress,
   initialized: false,
 };
 
@@ -47,9 +23,6 @@ export const commonSlice = createSlice({
       const address = state.addressList;
       state.addressList = address.filter((data) => data.id !== payload);
     },
-    setSelectAddress: (state, { payload }) => {
-      state.selectAddress = payload;
-    },
     setInitialized: (state) => {
       state.initialized = true;
     },
@@ -57,7 +30,7 @@ export const commonSlice = createSlice({
 });
 
 export const {
-  setCurrentAddress, deleteCurrentAddress, setSelectAddress, setInitialized,
+  setCurrentAddress, deleteCurrentAddress, setInitialized,
 } = commonSlice.actions;
 
 export default commonSlice.reducer;
